BoundaryIntegralEquations

Documentation for BoundaryIntegralEquations.

BoundaryIntegralEquations.AdjointDoubleLayer

BoundaryIntegralEquations.AdjointDoubleLayer

BoundaryIntegralEquations.AdjointDoubleLayer

BoundaryIntegralEquations.DiscreteClosedCurve

BoundaryIntegralEquations.DiscreteClosedCurve

BoundaryIntegralEquations.DiscreteClosedCurve

BoundaryIntegralEquations.DiscreteClosedCurve

BoundaryIntegralEquations.DoubleLayer

BoundaryIntegralEquations.DoubleLayer

BoundaryIntegralEquations.DoubleLayer

BoundaryIntegralEquations.Hypersingular

BoundaryIntegralEquations.Hypersingular

BoundaryIntegralEquations.SingleLayer

BoundaryIntegralEquations.SingleLayer

BoundaryIntegralEquations.SingleLayer

BoundaryIntegralEquations.ball

BoundaryIntegralEquations.cauchy_integral

BoundaryIntegralEquations.evaluate

BoundaryIntegralEquations.evaluate

BoundaryIntegralEquations.evaluate

BoundaryIntegralEquations.evaluate

BoundaryIntegralEquations.holomorphism_boundary_limit

BoundaryIntegralEquations.holomorphism_boundary_limit

BoundaryIntegralEquations.kernel

BoundaryIntegralEquations.kernel

BoundaryIntegralEquations.kernel

BoundaryIntegralEquations.kernel

BoundaryIntegralEquations.matrix

BoundaryIntegralEquations.periodic_spectral_diff

BoundaryIntegralEquations.populate_matrices!

BoundaryIntegralEquations.populate_matrices!

BoundaryIntegralEquations.solve

BoundaryIntegralEquations.solve

BoundaryIntegralEquations.starfish

BoundaryIntegralEquations.AdjointDoubleLayer — Type

AdjointDoubleLayer

Finite-dimensional approximation of the adjoint operator of a double-layer potential integral operator using the Nyström approach.

This operator maps a density on the boundary of a domain to the potential that it produces.

\[D^*: C^1(\Gamma) \to C^1(\bar \Omega) \\ D^*[u](x) = \int_\Gamma {\frac{\partial}{\partial n_x}k(x, y) \varphi(y) dS(y)}, x \in \Gamma, y \in \bar \Omega \]

BoundaryIntegralEquations.AdjointDoubleLayer — Method

AdjointDoubleLayer(equation::Laplace, source::AbstractManifold, target::AbstractMatrix, target_normals::AbstractMatrix, matrix_factory::Function=default_allocator)

Compute the operator coefficients for source-target interaction

Arguments

equation::Laplace: Partial Differential Equation corresponding to this operator

source::AbstractManifold: Manifold where the density is defined, i.e. points inside the domain

target::AbstractMatrix: Points where the potential is computed, i.e. points in the boundary of the domain

target_normals::AbstractMatrix: outward unit normal vectors at the target points

BoundaryIntegralEquations.AdjointDoubleLayer — Method

AdjointDoubleLayer(equation::Laplace, source::AbstractManifold, matrix_factory::Function=default_allocator)

Compute the operator coefficients for source self interaction

Arguments

equation::Laplace: Partial Differential Equation corresponding to this operator

source::AbstractManifold: Manifold where the density is defined and where the potential is computed, i.e. boundary of the domain

BoundaryIntegralEquations.DiscreteClosedCurve — Method

DiscreteClosedCurve(x::AbstractMatrix, v::AbstractMatrix, a::AbstractMatrix)

given positions, velocities, and accelerations of the curve parametrization, compute the remaining parameters

Arguments

x::AbstractMatrix: location of the points

v::AbstractMatrix: velocity of the curve at each point

a::AbstractMatrix: acceleration of the curve at each point

BoundaryIntegralEquations.DiscreteClosedCurve — Method

DiscreteClosedCurve(x::AbstractMatrix)

given positions, compute the velocities and accelerations using periodic spectral differentiation

Arguments

x::AbstractMatrix: locations of the points

BoundaryIntegralEquations.DiscreteClosedCurve — Method

DiscreteClosedCurve(θ::AbstractVector, ρ::Function)

construct curve given a list of parameter values and a parametrization

Arguments

θ::AbstractVector: list of nodes in parameter space

ρ::Function: function that parametrizes the curve

BoundaryIntegralEquations.DiscreteClosedCurve — Method

DiscreteClosedCurve(n_points::Int, ρ::Function)

construct curve given a number of points and a parametrization using equispaced nodes in parameter space

Arguments

n_points::Int: number of nodes

ρ::Function: function that parametrizes the curve

BoundaryIntegralEquations.DoubleLayer — Type

DoubleLayer

Finite-dimensional approximation of a double-layer potential integral operator using the Nyström approach.

This operator maps a density on the boundary of a domain to potential that it produces.

\[D: C^1(\Gamma) \to C^1(\bar \Omega)\\ D[u](x) = \int_\Gamma {\frac{\partial}{\partial n_y}k(x, y) \varphi(y) dS(y)}, x \in \bar \Omega, y \in \Gamma \]

BoundaryIntegralEquations.DoubleLayer — Method

DoubleLayer(equation::Laplace, source::AbstractManifold, # source manifold e.g. domain boundary, target::AbstractMatrix, # target points to compute operator, matrix_factory::Function=default_allocator)

Compute the operator coefficients for source-target interaction

Arguments

equation::Laplace: Partial Differential Equation corresponding to this operator

source::AbstractManifold: Manifold where the density is defined, i.e. boundary of the domain

target::AbstractMatrix: Points where the potential is computed

BoundaryIntegralEquations.DoubleLayer — Method

DoubleLayer(equation::Laplace, source::AbstractManifold, matrix_factory::Function=default_allocator)

Compute the operator coefficients for source self interaction

Arguments

equation::Laplace: Partial Differential Equation corresponding to this operator

source::AbstractManifold: Manifold where the density is defined and where the potential is computed, i.e. boundary of the domain

BoundaryIntegralEquations.Hypersingular — Type

Hypersingular

Finite-dimensional approximation of the hypersingular integral operator using the Nyström approach.

This operator maps a density on the boundary of a domain to the potential that it produces.

\[H: C^1(\Gamma) \to C^1(\Gamma)\\ H[u](x) = \int_\Gamma{\frac{\partial^2}{\partial n_x \partial n_y}k(x, y) \varphi(y) dS(y)}, x\in \Gamma, y \in \Gamma\]

BoundaryIntegralEquations.Hypersingular — Method

Hypersingular(equation::Laplace, source::AbstractManifold, correction::HypersingularCorrection, matrix_factory::Function=default_allocator)

Compute the operator coefficients for source self interaction

Arguments

equation::Laplace: Partial Differential Equation corresponding to this operator

source::AbstractManifold: Manifold where the density is defined and where the potential is computed, i.e. boundary of the domain

correction::HypersingularCorrection: Approach to handle the singularity

BoundaryIntegralEquations.SingleLayer — Type

SingleLayer

Finite-dimensional approximation of a single-layer potential integral operator using the Nyström approach.

This operator maps a density on the boundary of a domain to the potential that it produces.

\[S: C^1(\Gamma) \to C^1(\bar \Omega)\\ S[u](x) = \int_\Gamma {k(x, y) \varphi(y) dS(y)}, x \in \bar \Omega, y \in \Gamma \]

BoundaryIntegralEquations.SingleLayer — Method

SingleLayer(equation::Laplace, source::AbstractManifold, # source manifold e.g. domain boundary, target::AbstractMatrix, # target points to compute operator, matrix_factory::Function=default_allocator)

Compute the operator coefficients for source-target interaction

Arguments

equation::Laplace: Partial Differential Equation corresponding to this operator

source::AbstractManifold: Manifold where the density is defined, i.e. boundary of the domain

target::AbstractMatrix: Points where the potential is computed

BoundaryIntegralEquations.SingleLayer — Method

SingleLayer(equation::Laplace, source::AbstractManifold, correction::SingularCorrection, # order of kapur rokhlin singular correction, matrix_factory::Function=default_allocator)

Compute the operator coefficients for source self interaction

Arguments

equation::Laplace: Partial Differential Equation corresponding to this operator

source::AbstractManifold: Manifold where the density is defined and where the potential is computed, i.e. boundary of the domain

correction::SingularCorrection: Approach to deal with the singularity

BoundaryIntegralEquations.ball — Method

ball(r, n)

evenly distributed points in a circumference centered at the origin of radius r

Arguments

r: radius of the circumference

n: number of points

return: matrix with datapoints

BoundaryIntegralEquations.cauchy_integral — Method

cauchy_integral(target, source::DiscreteClosedCurve, boundary_data)

computes the inner value of a holomorphic function given its boundary data

Arguments

source::DiscreteClosedCurve: source manifold

target: target points

boundary_data: limit of the function at the curve

BoundaryIntegralEquations.evaluate — Function

evaluate(problem::BoundaryValueProblem{Laplace,Dirichlet,Interior,<:DiscreteClosedCurve}, approach::Indirect, correction::HypersingularCorrection, φ::BoundaryDensity, target::AbstractMatrix, relative_cutoff=0.05, matrix_factory::Function=default_allocator)

Evaluate the solution to a boundary value problem by using the solution to the associated boundary integral equation by the indirect approach.

Arguments

problem::BoundaryValueProblem{Laplace,Dirichlet,Interior,<:DiscreteClosedCurve}: problem whose solution is to be evaluated

approach::Indirect: the indirect approach

correction::HypersingularCorrection: type of correction to handle singular integral

φ::BoundaryDensity: the solution to the associated boundary integral equation

target::AbstractMatrix: points where the solution of the BVP needs evaluation

relative_cutoff: proportion of the characteristic length of the domain to switch between near-boundary evaluation using the Cauchy integral and conventional evaluation. Defaults to 5%.

BoundaryIntegralEquations.evaluate — Function

evaluate(problem::BoundaryValueProblem{Laplace,Dirichlet,Interior,<:DiscreteClosedCurve}, approach::Indirect, correction::HypersingularCorrection, φ::BoundaryDensity, target::AbstractMatrix, relative_cutoff=0.05, matrix_factory::Function=default_allocator)

compute Hypersingular integral operator and call evaluate

BoundaryIntegralEquations.evaluate — Method

evaluate(problem::BoundaryValueProblem{Laplace,Dirichlet,Interior,<:DiscreteClosedCurve}, approach::Direct, τ::Neumann, target::AbstractMatrix, ;, matrix_factory::Function=default_allocator)

evaluate the solution of a boundary value problem at requested locations, given the solution to the associated boundary integral equation through the direct approach, computing the required operators

Arguments

problem::BoundaryValueProblem{Laplace,Dirichlet,Interior,<:DiscreteClosedCurve}: boundary value problem to solve

approach::Direct: approach to solve the boundary integral equation

τ::Neumann: solution to the associated boundary integral equation

target::AbstractMatrix: locations to evaluate the solution to the BVP

BoundaryIntegralEquations.evaluate — Method

evaluate(problem::BoundaryValueProblem{Laplace,Dirichlet,Interior,<:DiscreteClosedCurve}, ::Direct, τ::Neumann, S_target::SingleLayer, D_target::DoubleLayer)

evaluate the solution of a boundary value problem at requested locations, given the solution to the associated boundary integral equation through the direct approach, and given precomputed operators

Arguments

problem::BoundaryValueProblem{Laplace,Dirichlet,Interior,<:DiscreteClosedCurve}: boundary value problem to solve

τ::Neumann: solution to the boundary integral equation through the direct approach

S_target::SingleLayer: source-target interaction integral operator

D_target::DoubleLayer: source-target interaction integral operator

BoundaryIntegralEquations.holomorphism_boundary_limit — Method

holomorphism_boundary_limit(problem<:BoundaryValueProblem{Laplace, Dirichlet, Interior}, density::BoundaryDensity)

computes the interior limit $v^-(x), x \in \Gamma$ of the holomorphic function $v(x), x \in \mathbb C \setminus \Gamma$, from the boundary density $\varphi$ which is the solution to the boundary integral equation by the indirect approach.

The The real part of $v$ is the the double-layer potential of the boundary density, i.e., it is the solution to the Dirichlet laplace problem.

This means, if $\varphi$ is the solution to

\[(D - \frac{1}{2})[\varphi] = \sigma \]

then the solution to the interior Laplace problem with Dirichlet data $\sigma$ is

\[u(x) = Re(v) = D[\varphi](x)\]

and since $v$ is holomorphic, it can be computed by a Cauchy integral, given the boundary limit of $v$.

\[v(x) = \frac{1}{2\pi i} \int_{\Gamma}{\frac{v^-(y)}{y - x} dy}\]

This function computes the boundary limit $v^-(x)$ from the boundary density $\varphi$ as:

\[v^-(x) = - \frac{1}{2} \varphi(x) - \frac{1}{2\pi i} \text{p.v.} \int_{\Gamma}{\frac{\varphi(y)}{y-x}dy} \]

Arguments

problem::BoundaryValueProblem{Laplace, Dirichlet, Interior, <:DiscreteClosedCurve}: Boundary value problem to solve

density::BoundaryDensity: Density that is the solution to the boundary integral equation associated to problem

BoundaryIntegralEquations.holomorphism_boundary_limit — Method

compute_boundary_limit(::Interior, op::SingleLayer{Laplace}, bc::BoundaryCondition, source::DiscreteClosedCurve)

[TODO:description]

Arguments

op::SingleLayer{Laplace}: [TODO:description]

bc::BoundaryCondition: [TODO:description]

source::DiscreteClosedCurve: [TODO:description]

BoundaryIntegralEquations.kernel — Method

kernel(::AdjointDoubleLayer{Laplace}, r_norm_sq, r_dot_nx)

Compute the kernel of the adjoint of the Laplace double-layer potential (DLP) integral operator, which is the normal derivative at x of the laplace SLP kernel.

\[\frac{\partial}{\partial n_x}k_{\text{Lap}}(x, y) = \nabla_x k_{\text{Lap}}(x, y) \cdot n_x = -\frac{1}{2\pi} \frac{(x - y) \cdot n_x}{|x - y|^2}\]

Arguments

r_norm_sq: square magnitude of the displacement vector

r_dot_nx: dot product between the displacement vector and the normal vector at x

BoundaryIntegralEquations.kernel — Method

kernel(::DoubleLayer{Laplace}, r_norm_sq, r_dot_ny)

Compute the kernel of the Laplace double-layer potential integral operator, which is the normal derivative at y of the laplace SLP kernel.

\[\frac{\partial}{\partial n_y}k_{\text{Lap}}(x, y) = \nabla_y k_{\text{Lap}}(x, y) \cdot n_y= \frac{1}{2\pi} \frac{(x - y) \cdot n_y}{|x - y|^2}\]

Arguments

r_norm_sq: square magnitude of the displacement vector

r_dot_ny: dot product between the displacement vector and the normal vector at y

BoundaryIntegralEquations.kernel — Method

kernel(::Hypersingular{Laplace}, r_norm_sq, r_dot_nx, r_dot_ny, nx_dot_ny)

Compute the kernel of the Laplace hypersingular integral operator, which is the mixed second-order normal derivative of the laplace SLP kernel.

\[\frac{\partial^2}{\partial n_x \partial n_y}k_{\text{Lap}}(x, y) = \frac{1}{2\pi} \left(\frac{-2((x - y) \cdot n_x) ((x - y) \cdot n_y)}{|x - y|^4} + \frac{n_x \cdot n_y}{|x - y|^2}\right)\]

Arguments

r_norm_sq: square magnitude of the displacement vector

r_dot_nx: dot product between the displacement vector and the normal vector at x

r_dot_ny: dot product between the displacement vector and the normal vector at y

nx_dot_ny: dot product between the normal vectory at x and the normal vector at y

BoundaryIntegralEquations.kernel — Method

kernel(::SingleLayer{Laplace}, r_norm_sq)

Compute the kernel of the Laplace single-layer potential (SLP) integral operator

\[k_{\text{Lap}}(x, y) = -\frac{1}{2\pi} \log|x - y|\]

Arguments

r_norm_sq: square magnitude of the displacement vector

BoundaryIntegralEquations.matrix — Method

matrix(op::IntegralOperator)

retrieve the matrix representation

Arguments

op::IntegralOperator: operator

BoundaryIntegralEquations.periodic_spectral_diff — Method

periodic_spectral_diff(d)

periodic spectral derivative

Arguments

f: matrix containing datapoints along curve

BoundaryIntegralEquations.populate_matrices! — Method

populate_matrices!(source::DiscreteClosedCurve{<:Real}, ops::IntegralOperator..., # variadic)

Compute the coefficients of several requested self-interaction integral operators

Arguments

source::DiscreteClosedCurve{<:Real}: Manifold where the density is defined

`ops::IntegralOperator... : List of operators that are required

BoundaryIntegralEquations.populate_matrices! — Method

populate_matrices!(source::DiscreteClosedCurve{<:Real}, ops, # tuple of operators with already allocated matrices)

Compute the coefficients of several requested source-target interaction integral operators

Arguments

source::DiscreteClosedCurve{<:Real}: Manifold where the density is defined

`ops::IntegralOperator... : List of operators that are required

target_normals: Outward unitary normal vectors at the target locations, in case the AdjointDoubleLayer operator is requested

BoundaryIntegralEquations.solve — Method

solve(problem::BoundaryValueProblem{Laplace,Dirichlet,Interior,<:DiscreteClosedCurve}, ::Direct, D_star::AdjointDoubleLayer, H::Hypersingular)

solve the boundary integral equation associated to the given boundary value problem using tthe direct approach by internally computing the required integral operators given precomputed integral operators

Arguments

problem::BoundaryValueProblem{Laplace,Dirichlet,Interior,<:DiscreteClosedCurve}: Boundary value problem to solve

D_star::AdjointDoubleLayer: Precomputed $D^*$ operator

H::Hypersingular: Precomputed $H$ operator

BoundaryIntegralEquations.solve — Method

solve(problem::BoundaryValueProblem{Laplace,Dirichlet,Interior,<:DiscreteClosedCurve}, approach::Direct, correction::HypersingularCorrection, ;, matrix_factory::Function=default_allocator)

solve the boundary integral equation associated to the given boundary value problem using tthe direct approach by internally computing the required integral operators

Arguments

problem::BoundaryValueProblem{Laplace,Dirichlet,Interior,<:DiscreteClosedCurve}: Problem to solve

approach::Direct: Solution approach

correction::HypersingularCorrection: Type of correction to use to deal with the singularity when computing the Hypersingular operator.

BoundaryIntegralEquations.starfish — Function

starfish(t, R=1, a=0.3, w=5)

starfish domain boundary in 2D as used by ref

parametrization of a Jordan curve that is a domain boundary

t |–> (x, y)

Arguments

t: [TODO:parameter]