Viscous and thermal losses

BoundaryIntegralEquations.LossyGlobalOuter — Type

LossyGlobalOuter(mesh::Mesh, freq;
            m=3,n=3,l=90,p=90,S=-1,sparsity=20.0,
            exterior=true,adaptive=false)

A LinearMap corresponding to the reduced lossy system.

BoundaryIntegralEquations.LossyGlobalInner — Type

LossyGlobalInner

A LinearMap corresponding to the inner system of the reduced lossy system.

BoundaryIntegralEquations.interpolation_function_derivatives — Function

interpolation_function_derivatives(mesh)

Returns interpolation function derivative matrics $(\mathbf{D}_x,\mathbf{D}_y,\mathbf{D}_z)$ of the 'physics_function'. This means that given the BEM interpolation of p the derivatives are given by

\[ \frac{\partial\mathbf{p}^\parallel}{\partial x} = \mathbf{D}_x\mathbf{p}, \newline \frac{\partial\mathbf{p}^\parallel}{\partial y} = \mathbf{D}_y\mathbf{p}, \newline \frac{\partial\mathbf{p}^\parallel}{\partial z} = \mathbf{D}_z\mathbf{p}.\]

NOTE! This is only the tangential part of the gradient. The reason being that the BEM interpolation on deal with surface values, and does therefore not contain the derivative information orthogonal to it.

BoundaryIntegralEquations.count_elements_pr_node — Function

count_elements_pr_node(mesh)

Counting the number of elements that each node is part of (Cornes can have various of numbers. Midpoints only 2.)