Experimental Features

BoundaryIntegralEquations.LossyOneVariableOuter — Type

LossyOneVariableOuter{T}

Struct representing the system matrix in the 1-variable system:

\[ \mathbf{A}\mathbf{p}_a = \mathbf{b}.\]

BoundaryIntegralEquations.LossyOneVariableInner — Type

LossyOneVariableInner{T}

Memory efficient representation of the matrix

\[(\mathbf{n}_1\mathbf{n}_1^\top + \mathbf{n}_2\mathbf{n}_2^\top + \mathbf{n}_3\mathbf{n}_3^\top)\circ(\mathbf{G}_v^{-1}\mathbf{H}_v) + (\mathbf{t}_1\mathbf{n}_1^\top + \mathbf{t}_2\mathbf{n}_2^\top + \mathbf{t}_3\mathbf{n}_3^\top)\circ\mathbf{D}_{t_1} + (\mathbf{s}_1\mathbf{n}_1^\top + \mathbf{s}_2\mathbf{n}_2^\top + \mathbf{s}_3\mathbf{n}_3^\top )\circ\mathbf{D}_{t_2},\]

which is part of the LossyOneVariableOuter{T}.

Multiplication with the struct is implemented and can therefore be used memory efficiently solve linear systems using an iterative solver.

BoundaryIntegralEquations.compute_lossy_rhs — Function

compute_lossy_rhs(A::LossyOneVariableOuter,vbn,vt1,vt2)

Computes the right-hand-side of the final 1-variable lossy system. Requires normal and tangential velocities.

BoundaryIntegralEquations.l_D_r! — Function

l_D_r!(D,lx,ly,lz,rx,ry,rz,tmp1,tmp2,y,x,setzero=true)

Computing the product

\[ ((\mathbf{l}_x\mathbf{r}_x^\top + \mathbf{l}_y\mathbf{r}_y^\top + \mathbf{l}_z\mathbf{r}_z^\top)\circ \mathbf{D})\mathbf{x}\]

Without performing the Hadamard product by applying the following three times

\[ ((\mathbf{l}\mathbf{r}^\top)\circ \mathbf{D})\mathbf{x} = (\text{diag}(\mathbf{l})\mathbf{D}\text{diag}(\mathbf{r}))\mathbf{x}\]

BoundaryIntegralEquations.l_GH_r! — Function

l_GH_r!(G,luG,H,lx,ly,lz,rx,ry,rz,tmp1,tmp2,y,x,setzero=true,uselu=false)

Computes the product

\[ ((\mathbf{l}_x\mathbf{r}_x^\top + \mathbf{l}_y\mathbf{r}_y^\top + \mathbf{l}_z\mathbf{r}_z^\top)\circ (\mathbf{G}^{-1}\mathbf{H}))\mathbf{x}\]

Without performing the Hadamard product by applying the following three times

\[ ((\mathbf{l}\mathbf{r}^\top)\circ (\mathbf{G}^{-1}\mathbf{H}))\mathbf{x} = (\text{diag}(\mathbf{l})\mathbf{G}^{-1}\mathbf{H}\text{diag}(\mathbf{r}))\mathbf{x}\]